Fix RA & declination errors (#5005).
All errors are now stored in degrees, so we need to convert them to arcsec as
appropriate.

Further, we no longer store "wm_ra_err" and "wm_decl_err" for runningcatalog
sources. Instead, we store wm_uncertainty_ew and wm_uncertainty_ns (both in
degrees). The latter can be used directly as the declination error; the former
needs inflating using the function convert.alpha() before it can be used as
the RA error.

Note that this introduces problems with some tables, since the errors
displayed in the table are no longer model fields. For now, I've just removed
the capability to sort on those columns. Presumably a little imagination on
the Django side could do a better job, but it's probably not worth the effort.