Show forcedfits_count on runcat_detail and varmetrics pages.
(Includes a fix  for failing test:

Fixture doesn't contain column for forcedfits_count. Generating
new fixture is quite some work. Settings
a default on the django side will set this in case of fixture
initisalisation, but won't do any harm when we recreate the
fixture in the future.)

add fields to correct page